Refactor user profile imports and improve component structure in Gantt header and dialogs
